An unreasonably huge commit
Sorry for that, I just happen to forget about the fact that git exists...
Can't remember all the features I've added, but among them:
Added bonk::AST structure that holds tree root and all its necessary buffers. Unlike bonk::TreeNode, bonk::AST is a self-sufficient structure
Implemented file importing through 'help' keyword, with implicit header files and caches. Tests should be added later.
Added such keywords as null, nothing, dabl, shrt, buul, long
Added casts, which are only used by compiler itself to generate constructors
Removed the x86 backend entirely
Added HIRFile and HIRLocation to make debug info. Only works for a patched QBE backend, and not on macOS (bruh).
Added a lot of tests, specifically for ARC, which turned out to be very buggy. Now it works fine, and I even managed to write a linked list, and bubble and merge sorting algorithms for it. Maybe I'll add them later as code examples.

+++ src/bonk/parsing/parser_position.hpp
@@ -9,9 +9,9 @@ namespace bonk {
 
 struct ParserPosition {
     std::string_view filename;
-    unsigned long line;
-    unsigned long ch;
-    unsigned long index;
+    unsigned int line;
+    unsigned int ch;
+    unsigned int index;
 
     // To print position
     friend std::ostream& operator<<(std::ostream& stream, const ParserPosition& position) {
